requirements.txt
Flask==3.0.3
numpy==2.1.3
pandas==2.2.3
pillow==11.0.0
opencv-python=4.12.0.88
